Nosetests now correctly fails on incorrect derivatives.
The problem was that a "%f" formatting was used on an AffineScalarFunc,
when generating the exception that indicates incorrect derivatives.
My nosetests did not report this as a problem and silently passed the
test.  I was not able to isolate and reproduce this behavior, though.

The new approach is that the numerical derivatives calculated through
uncertainties.NumericalDerivatives() are now given float arguments
instead of Variable arguments, in
test_uncertainties.py._compare_derivatives().  As a consequence,
the step of numerical derivatives is not anymore an AffineScalarFunc but
a simple float, as it should.  This makes the numerical derivative be
a float instead of an AffineScalarFunc, and the formatting works.