Test of __neg__, etc. now working again.
The problem was introduced by the new
test_uncertainties._compare_derivatives(), which forced derivatives to
be evaluated at float arguments.  This did not work for
AffineScalarFunc.__neg__ and co., which by definition do not apply to
floats.  A simple wrapping of float arguments to AffineScalarFunc before
evaluating the numerical derivatives, in the tests, solved the problem.