Automatic version numbering in setup.py.
Disadvantage of the new approach: it is more cumbersome to do `setup.py
register` in order to only update the package information on the
Python Package Index.  Before, it was possible to work on a new version
of the package (uncertainties.__version__ already set to a new version)
and update the package information of the previous uncertainties version
(as set in setup.py).