Logic of the formatting doc now simpler.
Instead of focusing on the "uncertainty control mode", the doc now
focuses on precision given/not given. This is better because the
uncertainty control mode is activated by "u", but one may wonder what
happens with '' or 'f' since there is no "u" in the format
specification. Concretely, users must simply remember that the
precision (given or not) is key to defining the formatting behavior
(explicit precision vs automatic precision). It is very reasonable
that .2f keeps the same meaning as before, so users who want a
precision defined as a number of digits can remember to use .2uf
instead. When they want something automatic, they just don't mention
any precision (no need to confuse users with the "u" modifier, here).