Fix for Python 2 (all tests pass now). Better comment.
The problem was that because of future, __slots__ = "new_attr" in
test_uncertainties.py is of some 'str' type, but isisntance(__slots__,
"str") is nonetheless false. This is a bit strange, but is likely
related to some tampering done with str. In any case, since Python 2
accepts unicode slot names (and they do work as expected), it was
actually a bug to not test the string character with basestring.