Drop official support for Python versions before 3.8
This change does not modify any code, so the codebase is still actually
compatible with Python 2.7 through 3.7, but it gives the project the
freedom to make incompatible changes with versions in that range in
subsequent commits.

A note in the documentation states that compatibility with earlier
versions could be considered if there is user demand.