Use setuptools-scm to ensure all repo files are in sdist
Listing `setuptools_scm` in `pyproject.toml` enables it to be used to
determine which files are included in the sdist instead of the older
`MANIFEST.in` procedure. `setuptools_scm` just includes every file
tracked by `git` which seems fine for this small project.

Additionally, delete `INSTALL.txt` which had outdated installation
information.