Fix readthedocs build (#261)
- [x] Closes #254 (insert issue number)
- [x] Executed `pre-commit run --all-files` with no errors
- [x] The change is fully covered by automated unit tests
- [x] Documented in docs/ as appropriate
- [x] Added an entry to the CHANGES file

Readthedocs are still not building. I would like to test this PR on RTD
before merging it in since it's taking a few tries to get this working.
It's possible to do this test on a side branch like this while keeping
the docs version hidden from the public.

@lebigot could you please make me and the other uncertainties
maintainers to be maintainers on the uncertainties readthedocs project?

I believe this can be done with something along the lines of
- Go to readthdocs account
- Go into the uncertainties package
- Go to "Admin"
- Go to tab for "Maintainers" 
- Add maintainers by user name. I can be added under the username
jagerber48. If you add me I would be happy to add the others.