Make codecov pull requests informational only (#269)
- [x] Closes #270 
- [x] Executed `pre-commit run --all-files` with no errors
- [x] The change is fully covered by automated unit tests
- [x] Documented in docs/ as appropriate
- [x] Added an entry to the CHANGES file

This PR hopes to resolve a few issues with `uncertainty` package
interaction with `codecov.io`.

- #203 introduced `.codecov.yaml`. Unfortunately, I think this was an
invalid configuration. Specifically because the `comment` section is
supposed to be a top-level section, not nested under the `codecov`
section. This means that since that PR the yaml configuration has NOT
been used for our codecov updates. I think this explains why we're
seeing codecov failures despite best attempts to suppress them.
- See for example #268 where a codecov failure is creating an X when we
really just want the PR to go through. This PR will (attempt to) make it
so that codecov coverage is reported, but will never result in a failed
coverage run. This way reviewers can see if patch coverage is < 100% or
if overall coverage drops due to a PR. If reviewers deem a lack or drop
of coverage to be significant they can block PR approval. Otherwise, for
minor coverage drops, the coverage can be easily ignored. It does so
using the `informational` flag in the `.codecov.yaml` file.
- It moves the API token back to using the github secret rather than
exposing it in the `.codecov.yaml` file.