Remove compare derivatives test function (#297)
- [x] Closes #296 (insert issue number)
- [x] Executed `pre-commit run --all-files` with no errors
- [x] The change is fully covered by automated unit tests
- [x] Documented in docs/ as appropriate
- [ ] Added an entry to the CHANGES file

Refactor the derivatives comparison tests to eliminate the
`compare_derivatives` test function in `helpers.py`. Right now the
complication of this test function is making addressing the topics in
#295 difficult.

This PR only impacts testing code. No code for actual usage is removed.
A `NumericalDerivatives` helper class was removed from `core.py` but
this class was only only used for testing and was not documented for
end-user usage.

Some `UFloat` methods and `umath` functions which were previously tested
for `UFloat`-calculated-derivative compared to numerical derivatives are
no longer tested in this way. The methods/functions for which testing
was dropped are listed below. These functions have been discussed in
#295 where there is at least tacit agreement that `uncertainties` can
drop these functions/methods entirely. So dropping their tests should be
uncontroversial.

The following `UFloat` methods are no longer tested in this way:

- `__abs__`
- `__trunc__`
- `__floordiv__`
- `__rfloordiv__`
- `__mod__`
- `__rmod__`

The following `umath` functions are no longer tested in this way:

- `copysign`
- `fabs`
- `fmod`
- `ldexp`
- `frac_part_modf`

---------

Co-authored-by: Matt Newville <newville@cars.uchicago.edu>