Refactor power derivative functions (#294)
- [x] Closes #293
- [x] Executed `pre-commit run --all-files` with no errors
- [x] The change is fully covered by automated unit tests
- [x] Documented in docs/ as appropriate
- [x] Added an entry to the CHANGES file

`uncertainties` allows calculating `x**y` where one or both of `x` and
`y` are `UFloat` objects. For this we need to calculate the `x` and `y`
derivatives of `x**y`. This PR refactors the functions that calculate
those derivatives to more explicitly and clearly handle non-typical
cases. Moving those functions to the module level also allows these
functions to be directly tested.

---------

Co-authored-by: Matt Newville <newville@cars.uchicago.edu>