Increase the stacklevel of the std_dev==0 warning (#305)
With this change, the warning will show the caller's location instead of
the location in `ufloat()` where the warning is generated.

I ran into some std_dev==0 warnings in a project I work on, partly from
lmfit (like reported
[here](https://github.com/lmfit/lmfit-py/issues/999)) and partly local
to the project (reporting parameter sets uniformly as `UFloat` whether
or not they have uncertainty). The `ufloat()` calls are made many levels
deep in the code so it was hard to tell where they were coming from (I
put a break point on the warning with a debugger). I think incrementing
the stack level so the caller is blamed for the warning is helpful.

- [ ] Closes # (insert issue number)
- [ ] Executed `pre-commit run --all-files` with no errors
- [ ] The change is fully covered by automated unit tests
- [ ] Documented in docs/ as appropriate
- [ ] Added an entry to the CHANGES file