Fix Sphinx copyright year substitution (#312)
Using a Unicode en dash to separate years in the Sphinx `copyright`
option results in unreproducible builds (in the sense of
https://reproducible-builds.org/), because Sphinx's special handling of
`SOURCE_DATE_EPOCH` only works when the separator is an ASCII hyphen.
The Sphinx developers seem to have considered changing this and then
decided against it (see
https://github.com/sphinx-doc/sphinx/pull/12738), so it seems best to
adjust the configuration here.

- [N/A] Closes # (insert issue number)
- [x] Executed `pre-commit run --all-files` with no errors
- [N/A] The change is fully covered by automated unit tests
- [N/A] Documented in docs/ as appropriate
- [x] Added an entry to the CHANGES file