Refactor `test_ufloat_from_str` (#320)
- [ ] Closes # (insert issue number)
- [ ] Executed `pre-commit run --all-files` with no errors
- [ ] The change is fully covered by automated unit tests
- [ ] Documented in docs/ as appropriate
- [ ] Added an entry to the CHANGES file

The motivation for this PR is that this is just one test that I happened
to refactor in the `4.x` branch. But the main refactor doesn't require
the core architecture changes, it was just a cleanup. It will be better
for the history if the main change appears on the `master` branch.
Consider this a part of #273.